I discovered an unfortunate effect of the change to lazily create all task entities; now it's very possible that systems who hold a task entity will not see its components until the next frame. This means that, given the original implementation of entity_is_complete, polling would return true before the task was finished. But I think this is the only way it can work in Legion, since entities can only be created within systems by the CommandBuffer (which does everything lazily, until the buffer is flushed). And I don't want to force users to flush the buffer after every system that creates tasks.
In order to keep the behavior the same between specs and Legion, I'm going to
keep the lazy task entity creation, and also change the data model a bit. In
this patch, I will make it so that task entities will never have their
TaskProgress component removed. This means that entity_is_complete can determine
that a task is not complete if it doesn't have a TaskProgress yet. It also means
that the TaskRunnerSystem cannot rely on a missing TaskProgress to skip over
the task, so instead it will check the is_complete() method. This incurs the
cost of additional AtomicBool reads.

I also considered adding another tag component that indicates whether a task
has finished being constructed. This seemed slightly more complex with only the
unmeasured benefit of removing AtomicBool reads from the TaskRunnerSystem.