bump version to 0.4.0
This includes several small, but nonetheless breaking, changes; the main one being
the OnCompletion::DeleteDescendents was removed, as it was associated with a bug in
polling for completion. It wasn't that useful to keep around anyway.

Also a bit of refactoring and renaming was done to make it clear the the TaskUser is
read-only and capable of making lazy task graph entity changes, while the TaskWriter
is capable of making such changes immediately; however, TaskGraph::assemble is not
capable of using the TaskWriter (yet); the hope is that eventually it will be easier
to create tasks that start running in the same frame they are created.